use records to pack axioms and data rather than Σ (reasons below)
When Σ-types are used, Agda is for some reason unable to infer a lot of
implicit parameters; this becomes too unwieldy when doing category
theory (e.g. natural transformations). I do not fully understand what it
is about Σ-types that causes this information-loss, but this solves the
problem.

In each case that I introduced a record, I proved it equivalent to a
Σ-type. This is necessary when characterizing path spaces.